update production fixture
