+++ GroupBundle/DataFixtures/MongoDB/LoadGroupDemoData.php
@@ -1,0 +1,36 @@
+<?php
+
+namespace OpenOrchestra\GroupBundle\DataFixtures\MongoDB;
+
+use Doctrine\Common\Persistence\ObjectManager;
+use OpenOrchestra\Backoffice\NavigationPanel\Strategies\AdministrationPanelStrategy;
+use OpenOrchestra\ModelInterface\DataFixtures\OrchestraProductionFixturesInterface;
+use OpenOrchestra\ModelInterface\DataFixtures\OrchestraFunctionalFixturesInterface;
+
+/**
+ * Class LoadGroupDemoData
+ */
+class LoadGroupDemoData extends AbstractLoadGroupData implements OrchestraProductionFixturesInterface, OrchestraFunctionalFixturesInterface
+{
+    /**
+     * @param ObjectManager $manager
+     */
+    public function load(ObjectManager $manager)
+    {
+        $group2 = $this->generateGroup('Demo group', 'Demo group', 'Groupe de démo', 'site2', 'group2');
+        $group2->addRole(AdministrationPanelStrategy::ROLE_ACCESS_REDIRECTION);
+        $manager->persist($group2);
+
+        $manager->flush();
+    }
+
+    /**
+     * Get the order of this fixture
+     *
+     * @return integer
+     */
+    public function getOrder()
+    {
+        return 610;
+    }
+}
